frontend/sensor_monitor_requirements.txt
python-dotenv
supabase
twilio
sqlalchemy
flask
requests
